+++ test/test_helpers/publishing_api_test.rb
@@ -7,35 +7,6 @@ describe GdsApi::TestHelpers::PublishingApi do
   let(:base_api_url) { Plek.current.find("publishing-api") }
   let(:publishing_api) { GdsApi::PublishingApi.new(base_api_url) }
 
-  describe "#assert_publishing_api_put_item" do
-    before { stub_default_publishing_api_put }
-    let(:base_path) { "/example" }
-
-    it "matches a put request with any empty attributes by default" do
-      publishing_api.put_content_item(base_path, {})
-      assert_publishing_api_put_item(base_path)
-    end
-
-    it "matches a put request with any arbitrary attributes by default" do
-      random_attributes = Hash[10.times.map {|n| [Random.rand * 1_000_000_000_000, Random.rand * 1_000_000_000_000]}]
-      publishing_api.put_content_item(base_path, random_attributes)
-      assert_publishing_api_put_item(base_path)
-    end
-
-    it "if attributes are specified, matches a request with at least those attributes" do
-      publishing_api.put_content_item(base_path, {"required_attribute" => 1, "extra_attrbibute" => 1})
-      assert_publishing_api_put_item(base_path, {"required_attribute" => 1})
-    end
-
-    it "matches using a custom request matcher" do
-      publishing_api.put_content_item(base_path, {})
-      matcher_was_called = false
-      matcher = ->(request) { matcher_was_called = true; true }
-      assert_publishing_api_put_item(base_path, matcher)
-      assert matcher_was_called, "matcher should have been called"
-    end
-  end
-
   describe '#request_json_matching predicate' do
     describe "nested required attribute" do
       let(:matcher) { request_json_matching({"a" => {"b" => 1}}) }